@@ -14,16 +14,19 @@
     "@fluentui/react": "^8.110.7",
     "@fluentui/react-components": "^9.7.2",
     "@fluentui/react-icons": "2.0.226",
-    "@ms-fabric/workload-client": "^2.0.0",
+    "@ms-fabric/workload-client": "^2.0.4",
     "history": "^4.9.0",
     "jwt-decode": "^3.1.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "i18next": "^23.12.3",
+    "i18next-http-backend": "^2.4.2",
+    "react-i18next": "^15.0.1",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-router-dom": "^5.3.4"
   },
   "devDependencies": {
-    "@types/react": "^17.0.36",
-    "@types/react-dom": "^17.0.11",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
     "@types/react-router-dom": "^5.3.2",
     "clean-webpack-plugin": "^4.0.0",
     "copy-webpack-plugin": "^10.0.0",
